debug

+++ code_app.py
@@ -368,7 +368,7 @@ with st.sidebar:
                 
                 labels_df = spot_600[(spot_600.release_date >= begin_date) & (spot_600.release_date <= end_date)].copy()
                 labels_df = labels_df.groupby(['artist_1', 'artist_2']).agg(num_feats=('track_id', 'count')).reset_index()
-                labels_df['done_feat'] = (labels_df.num_feats >= 1).int()
+                labels_df['done_feat'] = (labels_df.num_feats >= 1)
                 st.markdown(labels_df.head())
                 df_select = pd.merge(df_select, labels_df[['artist_1', 'artist_2', 'done_feat']],
                                     on=['artist_1', 'artist_2'],
